parameter pollution workinggit status

+++ implementation1/email.py
@@ -36,38 +36,36 @@ def generate_6_digit_OTP(cur):
 
 if __name__== "__main__":
     POST={}
-    args=sys.argv[1].split('&')
+    args=str(sys.argv[1]).split('&')
 
     for arg in args:
         t=arg.split('=')
         if len(t)>1 and t[0] not in POST: 
             POST[t[0]] = t[1]
-            k, v=arg.split('='); POST[k]=v
 
-    email = POST.get('email')
+    email = POST["email"]
 
     email = urllib.parse.unquote(email)
-
-    # print(email)
-
-
-    # email = ""
-    # if len(sys.argv) >= 2:
-    #     email = sys.argv[1]
 
     conn = connect()
     cur = conn.cursor()
 
-    OTP = generate_6_digit_OTP(cur)
+    #make sure there's an account for this email
+    validate_acct = cur.execute("SELECT * FROM credentials WHERE email = %s", [email])
+    acct_rows = cur.fetchall()
+    acct_rows_num = len(acct_rows)
 
-    # email = POST.get("email")
+    if acct_rows_num > 0:
+        OTP = generate_6_digit_OTP(cur)
 
-    cur.execute("DELETE FROM otp WHERE account=%s", [email])
+        #if there are already an OTPs for this account, delete them
+        cur.execute("DELETE FROM otp WHERE account=%s", [email])
 
-    cur.execute("INSERT INTO otp (code, account) VALUES (%s, %s)", [OTP, email])
+        #insert the OTP and email into the database
+        cur.execute("INSERT INTO otp (code, account) VALUES (%s, %s)", [OTP, email])
     conn.commit()
     if cur:
         cur.close()
     if conn:
         conn.close()
-    print(email)
+    print(email,end='')
